Refactoring

+++ tests/unit_tests/test_task.py
@@ -9,21 +9,22 @@ def test_create_task(client, session, daily_task):
     assert session.query(exists().where(Task.id == 1)).scalar()
 
 
-def test_delete_task(client, session, daily_task):
-    response = client.delete(url="api/task/1")
+def test_delete_task(client, session, daily_task, token):
+    response = client.delete(url="api/task/1", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_200_OK
     assert session.query(func.count(Task.description)).scalar() == 0
 
 
-def test_update_task(client, session, daily_task):
+def test_update_task(client, session, daily_task, token):
     updated_task = {
+        "user_id": 1,
         "status": "Создано",
         "end_date": "2022-12-27T12:10:03.500Z",
         "description": "Пойти в ресторан",
-        "category": {"name": "Основные"},
+        "category": "Основные",
         "priority": "Срочно сделать",
     }
-    response = client.put("/api/task/1", json=updated_task)
+    response = client.put("/api/task/1", json=updated_task, headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_200_OK
     assert session.query(
         exists().where(Task.description == updated_task.get("description"))
@@ -31,9 +32,8 @@ def test_update_task(client, session, daily_task):
 
 
 def test_get_active_tasks_being_authorized(
-    client, daily_task, authorized_user
+    client, daily_task, token
 ):
-    token = authorized_user["access_token"]
     response = client.get(
         "/api/tasks/active/1", headers={"Authorization": f"Bearer {token}"}
     )
@@ -50,8 +50,7 @@ def test_get_active_tasks_being_unauthorized(client, daily_task):
     assert "Invalid token" in response.json()["detail"]
 
 
-def test_get_done_tasks_being_authorized(client, done_task, authorized_user):
-    token = authorized_user["access_token"]
+def test_get_done_tasks_being_authorized(client, done_task, token):
     response = client.get(
         "/api/tasks/done/1", headers={"Authorization": f"Bearer {token}"}
     )
